requirements.txt
python-telegram-bot
instaloader
opencv-python-headless
numpy
python-dotenv
yt-dlp
flask
gunicorn

